+++ imports/message-read-receipt/client/index.js
@@ -1,1 +1,4 @@
+import './main';
+import './message';
 import './readReceipts';
+import './room';
