+++ imports/message-read-receipt/client/main.js
@@ -1,0 +1,7 @@
+Template.main.helpers({
+	readReceiptsEnabled() {
+		if (RocketChat.settings.get('Message_Read_Receipt_Store_Users')) {
+			return 'read-receipts-enabled';
+		}
+	}
+});
