+++ packages/rocketchat-ui-master/client/main.html
@@ -48,7 +48,7 @@
 						{{#unless modal}}
 							{{> sideNav }}
 						{{/unless}}
-						<div class="rc-old main-content content-background-color {{#if modal}}main-modal{{/if}}">
+						<div class="rc-old main-content content-background-color {{readReceiptsEnabled}} {{#if modal}}main-modal{{/if}}">
 							{{> Template.dynamic template=center}}
 						</div>
 					</div>
